requirements.txt
pydicom
numpy
Pillow
transformers>=4.50.0
torch
accelerate
bitsandbytes
scipy
fastapi
uvicorn
python-multipart
streamlit
requests
huggingface_hub
